Basic POI support
The LBL file contains a section for detailed information
about POIs (street, streetnumber, ZIP, phone number). These
entries can be linked from the Point-entries in the map
subdivisions.

For now we only use the Name field in the POIRecord.
The user visible change is in MapSource: Most POIs now have
a Property context-menu entry, which shows the detailed
info.

The Point-entry flag was found in qlandkarte. The rest is
documented in the imgformat pdf.

- Elrond

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@581 25d90789-57f7-4ee0-8453-03a3dfeeeb22