Lots of changes mostly just to make the StyledConverter work in a backward compatible way with existing map-features.csv files.
New feature as suggested by Elrond:  you can give a default name
for each osm rule, that is used when there is no name.  This is useful in
areas that are not completely mapped as many osm features have to use the
same garmin symbol and if they don't have a name it might be difficult to
know what they are tagged as.



git-svn-id: http://svn.parabola.me.uk/mkgmap/branches/style@591 25d90789-57f7-4ee0-8453-03a3dfeeeb22