Merged the MDR code back to trunk
One change to compare labels ignoring case was required to produce identical
img files.  Plus unrelated test failure from when the size of the NOD section was reduced
fixed.

Note that the default overview-mapname is now osmmap rather than 63240000.

The --index option creates a global index file, named by default osmmap_mdr.img along with
an osmmap.mdx file.

Although this should not have affected any existing feature, it is a rather large change so please
be on the lookout for any problems.

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@1283 25d90789-57f7-4ee0-8453-03a3dfeeeb22