Added --report-dead-ends option to warn about roads that go nowhere.
Can generate two different warning messages:

"Oneway street goes nowhere" and "Confluence of oneway streets".

The first is output for oneway roads that don't join another routable way
this occurs quite a lot (i.e. roads into car parks, etc.) so a lot of
those warnings can be ignored.

The second is output when multiple oneway streets meet with nowhere else
to go. This is very likely to be an error in the map data (i.e. motorway
segment in the wrong direction) and worthy of investigation.

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@1301 25d90789-57f7-4ee0-8453-03a3dfeeeb22