Disable test that checks if lines and shapes are within the bounds.
This test catches a few shapes whose start point is outside of the
area's bounds. But as not all of the shapes whose start point is outside
of the area will actually have a bad delta lat/lon, we may as well let
them all through and just suffer the odd failure later.

Sorry, I just don't have the capacity at the moment to work on a real
solution to this thorny problem.

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@1392 25d90789-57f7-4ee0-8453-03a3dfeeeb22